Make WIFI driver loading mechanism configurable
Currently, on Samsung platform, it loads WIFI driver dynamically,
but on Intel platform, WIFI driver is loaded when system boots up.
So it needs to make the WIFI driver loading mechanism configurable
to adapt both Samsung and Intel platform.

Use the existing macro definition "EMBEDDED_TARGET" to fix this issue.

Change-Id: Iddccfc411f79284232f9d10b971ac6cd28ae75ca
Signed-off-by: Chengyi Zhao <chengyi1.zhao@archermind.com>